@@ -20,14 +20,14 @@
     "queue-interop/queue-interop": "^0.8.1",
     "bunny/bunny": "^0.5.5",
     "php-amqplib/php-amqplib": "^3.1",
-    "doctrine/dbal": "^3.2",
     "ramsey/uuid": "^4.3",
     "psr/log": "^1.1 || ^2.0 || ^3.0",
     "psr/container": "^1.1 || ^2.0",
     "makasim/temp-file": "^0.2",
     "google/cloud-pubsub": "^1.46",
-    "doctrine/orm": "^2.12",
-    "doctrine/persistence": "^2.0|^3.0",
+    "doctrine/dbal": "^3.6.2|^4.0",
+    "doctrine/orm": "^2.20|^3.5",
+    "doctrine/persistence": "^3.4|^4.1",
     "mongodb/mongodb": "^1.17",
     "pda/pheanstalk": "^3.1",
     "aws/aws-sdk-php": "^3.290",
@@ -38,10 +38,12 @@
     "predis/predis": "^2.1",
     "thruway/client": "^0.5.5",
     "thruway/pawl-transport": "^0.5.2",
+    "ratchet/pawl": "^0.4.3",
+    "ratchet/rfc6455": "^0.4.0",
     "influxdb/influxdb-php": "^1.14",
     "datadog/php-datadogstatsd": "^1.3",
     "guzzlehttp/guzzle": "^7.0.1",
-    "guzzlehttp/psr7": "^1.9.1",
+    "guzzlehttp/psr7": "^2.4.5",
     "php-http/discovery": "^1.13",
     "voryx/thruway-common": "^1.0.1",
     "react/dns": "^1.4",
@@ -53,20 +55,20 @@
     "phpunit/phpunit": "^9.5.28",
     "phpstan/phpstan": "^1.0",
     "queue-interop/queue-spec": "^0.6.2",
-    "symfony/browser-kit": "^6.2|^7.0",
-    "symfony/config": "^6.2|^7.0",
-    "symfony/process": "^6.2|^7.0",
-    "symfony/console": "^6.2|^7.0",
-    "symfony/dependency-injection": "^6.2|^7.0",
-    "symfony/event-dispatcher": "^6.2|^7.0",
-    "symfony/expression-language": "^6.2|^7.0",
-    "symfony/http-kernel": "^6.2|^7.0",
-    "symfony/filesystem": "^6.2|^7.0",
-    "symfony/framework-bundle": "^6.2|^7.0",
-    "symfony/validator": "^6.2|^7.0",
-    "symfony/yaml": "^6.2|^7.0",
+    "symfony/browser-kit": "^7.0|^8.0",
+    "symfony/config": "^7.0|^8.0",
+    "symfony/process": "^7.0|^8.0",
+    "symfony/console": "^7.0|^8.0",
+    "symfony/dependency-injection": "^7.0|^8.0",
+    "symfony/event-dispatcher": "^7.0|^8.0",
+    "symfony/expression-language": "^7.0|^8.0",
+    "symfony/http-kernel": "^7.0|^8.0",
+    "symfony/filesystem": "^7.0|^8.0",
+    "symfony/framework-bundle": "^7.0|^8.0",
+    "symfony/validator": "^7.0|^8.0",
+    "symfony/yaml": "^7.0|^8.0",
     "empi89/php-amqp-stubs": "*@dev",
-    "doctrine/doctrine-bundle": "^2.5",
+    "doctrine/doctrine-bundle": "^2.5|^3.1",
     "doctrine/mongodb-odm-bundle": "^4.7|^5.0",
     "alcaeus/mongo-php-adapter": "^1.0",
     "kwn/php-rdkafka-stubs": "^2.0.3",
@@ -127,7 +129,7 @@
       "ext-rdkafka": "4.0",
       "ext-bcmath": "1",
       "ext-mbstring": "1",
-      "ext-mongodb": "1.17.3",
+      "ext-mongodb": "1.21",
       "ext-sockets": "1"
     },
     "prefer-stable": true,
