@@ -9,15 +9,15 @@
         "php": "^8.1",
         "enqueue/enqueue": "^0.10",
         "queue-interop/queue-interop": "^0.8",
-        "symfony/event-dispatcher": "^5.4|^6.0"
+        "symfony/event-dispatcher": "^7.0|^8.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
-        "symfony/dependency-injection": "^5.4|^6.0",
-        "symfony/config": "^5.4|^6.0",
-        "symfony/http-kernel": "^5.4|^6.0",
-        "symfony/filesystem": "^5.4|^6.0",
-        "symfony/yaml": "^5.4|^6.0",
+        "symfony/dependency-injection": "^7.0|^8.0",
+        "symfony/config": "^7.0|^8.0",
+        "symfony/http-kernel": "^7.0|^8.0",
+        "symfony/filesystem": "^7.0|^8.0",
+        "symfony/yaml": "^7.0|^8.0",
         "enqueue/null": "0.10.x-dev",
         "enqueue/fs": "0.10.x-dev",
         "enqueue/test": "0.10.x-dev"
@@ -30,7 +30,7 @@
         "docs": "https://github.com/php-enqueue/enqueue-dev/blob/master/docs/index.md"
     },
     "suggest": {
-        "symfony/dependency-injection": "^5.4|^6.0 If you'd like to use async event dispatcher container extension."
+        "symfony/dependency-injection": "^7.0|^8.0 If you'd like to use async event dispatcher container extension."
     },
     "autoload": {
         "psr-4": { "Enqueue\\AsyncEventDispatcher\\": "" },
