@@ -7,7 +7,7 @@
     "license": "MIT",
     "require": {
         "php": "^8.1",
-        "symfony/framework-bundle": "^6.2|^7.0",
+        "symfony/framework-bundle": "^7.0|^8.0",
         "queue-interop/amqp-interop": "^0.8.2",
         "queue-interop/queue-interop": "^0.8",
         "enqueue/enqueue": "^0.10",
@@ -36,13 +36,13 @@
         "enqueue/async-event-dispatcher": "0.10.x-dev",
         "enqueue/async-command": "0.10.x-dev",
         "php-amqplib/php-amqplib": "^3.0",
-        "doctrine/doctrine-bundle": "^2.3.2",
+        "doctrine/doctrine-bundle": "^2.5|^3.1",
         "doctrine/mongodb-odm-bundle": "^3.5|^4.3|^5.0",
         "alcaeus/mongo-php-adapter": "^1.0",
-        "symfony/browser-kit": "^6.2|^7.0",
-        "symfony/expression-language": "^6.2|^7.0",
-        "symfony/validator": "^6.2|^7.0",
-        "symfony/yaml": "^6.2|^7.0"
+        "symfony/browser-kit": "^7.0|^8.0",
+        "symfony/expression-language": "^7.0|^8.0",
+        "symfony/validator": "^7.0|^8.0",
+        "symfony/yaml": "^7.0|^8.0"
     },
     "suggest": {
         "enqueue/async-command": "If want to run Symfony command via message queue",
