@@ -17,12 +17,12 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
-        "symfony/console": "^5.41|^6.0",
-        "symfony/dependency-injection": "^5.4|^6.0",
-        "symfony/config": "^5.4|^6.0",
-        "symfony/event-dispatcher": "^5.4|^6.0",
-        "symfony/http-kernel": "^5.4|^6.0",
-        "symfony/yaml": "^5.4|^6.0",
+        "symfony/console": "^7.0|^8.0",
+        "symfony/dependency-injection": "^7.0|^8.0",
+        "symfony/config": "^7.0|^8.0",
+        "symfony/event-dispatcher": "^7.0|^8.0",
+        "symfony/http-kernel": "^7.0|^8.0",
+        "symfony/yaml": "^7.0|^8.0",
         "enqueue/amqp-ext": "0.10.x-dev",
         "enqueue/amqp-lib": "0.10.x-dev",
         "enqueue/amqp-bunny": "0.10.x-dev",
@@ -42,9 +42,9 @@
         "enqueue/dsn": "0.10.x-dev"
     },
     "suggest": {
-        "symfony/console": "^5.4|^6.0 If you want to use cli commands",
-        "symfony/dependency-injection": "^5.4|^6.0",
-        "symfony/config": "^5.4|^6.0",
+        "symfony/console": "^7.0|^8.0 If you want to use cli commands",
+        "symfony/dependency-injection": "^7.0|^8.0",
+        "symfony/config": "^7.0|^8.0",
         "enqueue/amqp-ext": "AMQP transport (based on php extension)",
         "enqueue/stomp": "STOMP transport",
         "enqueue/fs": "Filesystem transport",
