@@ -11,16 +11,16 @@
         "enqueue/null": "^0.10",
         "queue-interop/queue-interop": "^0.8",
         "doctrine/orm": "^2.12",
-        "doctrine/dbal": "^2.12 | ^3.0"
+        "doctrine/dbal": "^3.2|^4.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
         "enqueue/test": "0.10.x-dev",
-        "doctrine/doctrine-bundle": "^2.3.2",
-        "symfony/browser-kit": "^5.4|^6.0",
-        "symfony/expression-language": "^5.4|^6.0",
-        "symfony/framework-bundle": "^5.4|^6.0",
-        "symfony/yaml": "^5.4|^6.0"
+        "doctrine/doctrine-bundle": "^2.5|^3.1",
+        "symfony/browser-kit": "^7.0|^8.0",
+        "symfony/expression-language": "^7.0|^8.0",
+        "symfony/framework-bundle": "^7.0|^8.0",
+        "symfony/yaml": "^7.0|^8.0"
     },
     "support": {
         "email": "opensource@forma-pro.com",
