@@ -9,7 +9,7 @@ use Symfony\Component\DependencyInjection\Extension\Extension;
 
 class AsyncCommandExtension extends Extension
 {
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         foreach ($configs['clients'] as $client) {
             // BC compatibility
