@@ -13,7 +13,7 @@ use Symfony\Component\DependencyInjection\Reference;
 
 class AsyncEventDispatcherExtension extends Extension
 {
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $config = $this->processConfiguration(new Configuration(), $configs);
 
