@@ -41,10 +41,19 @@ class DoctrinePingConnectionExtensionTest extends TestCase
             ->expects($this->never())
             ->method('close')
         ;
-        $connection
-            ->expects($this->never())
-            ->method('connect')
-        ;
+        if (
+            method_exists(Connection::class, 'connect')
+            && (new \ReflectionMethod(Connection::class, 'connect'))->isPublic()
+        ) {
+            // DBAL < 4
+            $connection->expects($this->never())
+                ->method('connect');
+        } else {
+            // DBAL >= 4, calls connect() internally
+            $connection
+                ->expects($this->never())
+                ->method('getServerVersion');
+        }
 
         $context = $this->createContext();
 
@@ -81,10 +90,18 @@ class DoctrinePingConnectionExtensionTest extends TestCase
             ->expects($this->once())
             ->method('close')
         ;
-        $connection
-            ->expects($this->once())
-            ->method('connect')
-        ;
+        if (
+            method_exists(Connection::class, 'connect')
+            && (new \ReflectionMethod(Connection::class, 'connect'))->isPublic()
+        ) {
+            // DBAL < 4
+            $connection->expects($this->once())
+                ->method('connect');
+        } else {
+            // DBAL >= 4, calls connect() internally
+            $connection->expects($this->once())
+                ->method('getServerVersion');
+        }
 
         $context = $this->createContext();
 
