@@ -12,7 +12,7 @@ use Interop\Queue\Processor;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\HttpKernel\DependencyInjection\ServicesResetter;
+use Symfony\Contracts\Service\ResetInterface;
 
 class ResetServicesExtensionTest extends TestCase
 {
@@ -50,8 +50,8 @@ class ResetServicesExtensionTest extends TestCase
     /**
      * @return MockObject|ManagerRegistry
      */
-    protected function createResetterMock(): ServicesResetter
+    protected function createResetterMock(): ResetInterface
     {
-        return $this->createMock(ServicesResetter::class);
+        return $this->createMock(ResetInterface::class);
     }
 }
