@@ -24,7 +24,7 @@ class QueueConsumerOptionsCommand extends Command
         $this->consumer = $consumer;
     }
 
-    protected function configure()
+    protected function configure(): void
     {
         parent::configure();
 
